Optimize SOSCF performance
